Extract workspace attention flash ring view
