+++ Sources/Panels/WorkspaceAttentionFlashRingView.swift
@@ -1,0 +1,20 @@
+import SwiftUI
+
+struct WorkspaceAttentionFlashRingView: View {
+    let opacity: Double
+    var reason: WorkspaceAttentionFlashReason = .navigation
+
+    var body: some View {
+        let presentation = WorkspaceAttentionCoordinator.flashStyle(for: reason)
+        let color = Color(nsColor: presentation.accent.strokeColor)
+
+        RoundedRectangle(cornerRadius: FocusFlashPattern.ringCornerRadius)
+            .stroke(color.opacity(opacity), lineWidth: PanelOverlayRingMetrics.lineWidth)
+            .shadow(
+                color: color.opacity(opacity * presentation.glowOpacity),
+                radius: presentation.glowRadius
+            )
+            .padding(FocusFlashPattern.ringInset)
+            .allowsHitTesting(false)
+    }
+}
